ai-nlp/requirements.txt
spacy>=3.7.0
PyPDF2>=3.0.0
python-docx>=1.1.0
docx2txt>=0.8
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
google-generativeai>=0.8.0
